feat(importmap): add Flowbite Library
This was a bit tricky to get working, mainly due to the CSS. I tried importing the css from the ./application/stylesheets/application.css file with @import flowbite per a guide I was reading. This worked but not completely, when I used the tooltip the arrow portion wasn't rendering. I switched to linking directly to the stylesheet in the application/layouts.html.erb file.