feat(dependency): disbale flowbite.js
The current functionality is not operational when triggered by link_to clicks. This issue may stem from a potential misconfiguration on my part in managing the Flowbite dependency. I plan to revisit and address this concern once I integrate company services into the project. It's worth noting that this feature was initially considered a 'nice-to-have' and is not currently deemed a necessity at the moment.