feat(models): Add Review; Configure Associations
I decided to go forward with modeling Reviews as a polymorphic association. At the moment, there are no other reviewable things, but I foresee the need to review other things or rating things in the future, such as quizzes. It would be a bit of work to remodel this later, so I think it's worth doing this now.