fix(js): TypeError failed to resolve module flowbite
Git commit c21e49e8... removed JS dependency Flowbite as it was breaking some CSS on the site. It will be added back once the feature that needs its supported is reintroduced. The import specifier wasn't resolving this was causing JS issues across the app, for example devise controller actions, and chartkick working on the quiz results page.