feat(wip): allow users to create quizzes
This is the first step to allowing users to create
quizzes. The first step is defining a quiz that will start in
draft mode. Next step is allowing them to build the quiz with
questions, and finally giving them an option to publish as 'private' for
their own use or as 'public' to share with others.