feat: init start/end times on new driver logs
Driver logs will initialize to the beginning of the day and end of the
day. A log represents a drivers duty status between a 24 hour period. In
the future the driver will be able to record backlog records where they
specify a day. The end and start times will still be set by the
application.